Sokoban: simplify levels database code
