+++ sokoban/scripts/app.h
@@ -15,6 +15,7 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public
 
 #pragma once
 #include <gui/gui.h>
+#include "levels_database.h"
 
 typedef struct SceneManager SceneManager;
 typedef struct AppGameplayState AppGameplayState;
@@ -41,22 +42,3 @@ typedef struct AppGameplayState
     int selectedCollection;
     int selectedLevel;
 } AppGameplayState;
-
-typedef struct LevelItem
-{
-    unsigned short worldBest;
-    unsigned short playerBest;
-} LevelItem;
-
-typedef struct LevelsCollection
-{
-    char name[32];
-    int levelsCount;
-    LevelItem* levels;
-} LevelsCollection;
-
-typedef struct LevelsDatabase
-{
-    int collectionsCount;
-    LevelsCollection* collections;
-} LevelsDatabase;
